refactor(linter): Move go linters to new file
